+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.5.1
 * Fix a bug where macros were not properly disabled under `NDEBUG`
+* Code cleanup and consolidation
 
 # 0.5.0
 * Support for all fundamental integer and float types